rename tree to _base.py
